requirements.txt
# Core ML & Computer Vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
albumentations>=1.3.0

# Data & Dataset Management
kagglehub>=0.2.0

# Inference & API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
onnx>=1.14.0
onnxruntime>=1.15.0

# Monitoring & Logging
prometheus-client>=0.17.0
structlog>=23.1.0
python-json-logger>=2.0.7

# Utilities
pyyaml>=6.0
click>=8.1.0
python-dotenv>=1.0.0
requests>=2.31.0
boto3>=1.28.0
google-cloud-storage>=2.10.0
